nifga deprecation of all terms
This is the main NIF-GrossAnatomy deprecation commit.
All classes in the old NIFGA namespace have either been
replacedBy: an UBERON equivalent or they oboInOwl:hasDbXref
to a deprecated UBERON id or they are replaced by nothing
(there are very few nothings, and most were already deprecated).
The code that oversaw this is in tgbugs/pyontutils nifga_deprecation.py.
There is a commit following this one that ports the existing
ro:has_proper_part and ro:proper_part_of partonomy to UBERON
replacements. Members of the hpp tree are all present in uberon.