NIFORG part of issues #70 hasTaxonRank
Turns out that the issues with hasTaxonRank were the result of an 'off
by 6310' error, and a reminder why using annotation properties to hold
identifiers isn't alwasy the best idea. Not entirely sure how the issue
came about (some script running a regex that hit stuff out of scope?)
but it has been there for years. Now resolved.