normalization of annotaiton properties
Over time various kinds of copy paste errors have led to annotation
properties defined in OBOANN and BIRNANN to be placed in the wrong
namespace. This commit corrects those issues and also removes a couple
of deprecated classes. All files modified were also reserialized so that
their ordering is correct.