NIFCHEM NIFMOL chebi-bridge import chain
All chemical and molecular entities are now imported via chebi-bridge.
We may want to change this one more time so that the merge between
NIF-Chemical and chebi-bridge is done in chemical-bridge so that it is
more obvious, but since chebi-bridge makes such heavy use of NIFCHEM
terms it seems to make more sense this way. This commit also resolves
the circular import issues with NIF-Molecule. Relevant to:
https://github.com/SciCrunch/NIF-Ontology/issues/115