anatomy-bridge, uberon, parcellation import chain
Major update to the way anatomy entities are imported into NIFSTD>
Everything (almost) is not routed through anatomy-bridge.ttl which will
make it much cleaner and easier to manage merging all the different
files that will be needed for mapping parcellation schemes to Uberon.
Note that uberon.owl is now imported via uberon-bridge.ttl to match
behavior of doid-bridge.ttl and chebi-bridge.ttl.