NIFSTD UBERON reconciliation part 2
Switched subClassOf Regional Part of to proper_part_of for membranes
which were subclassed to things that are now organs. This mirrors the
modelling in uberon for skirting issues with acellular vs multicellular
anatomical structures. Some addional changes in uberon-bridge are not
real changes but reordering because I previously forgot to run ttlfmt on
the file.

See aslo https://github.com/obophenotype/uberon/issues/1358