removal of old prefixes using qnamefix
Result of the execution of qnamefix.py (in ttl/) from pyontutils commit
53f5e7a41b8500fe4639d860f1d2419fe5048492.

This commit completes the uri switch to NIFSTD. The only remaining
iris that *could* be prefixed using the old prefixes are in annotation
properties and have special cases. Below are the two oneliners I used to
verify that the switch has gone according to plan.

review of all prefixes
`grep -r "@prefix" | grep -v NIF-NIFSTD | grep -v Binary | \
 cut -d':' -f2- | sort -u`

review of remaining unprefixed urls
`grep "<http.\+>" *.ttl */*.ttl */*/*.ttl | grep -v 'NIF-NIFSTD' | \
 grep -v '\.ttl>' | grep -v '\.owl>' | grep -v "@prefix" | \
 cut -d'<' -f2 | cut -d'>' -f1 | sort -u`

This commit also completes the ontology side of
https://github.com/SciCrunch/NIF-Ontology/issues/59. However we still
need to make sure that all the new uris in NIF-NIFSTD-mapping.ttl work
with the resolver and ultimately resolve to an interlex record (which
will come down the road as soon as we can switch off the old neurolex).