added uberon-dead, so-dead, dead improvements
This commit generally completes the managment of deprecated classes
coming from external ontologies. uberon-dead was added to the import
chain, but so-dead was not since I need to rework the chemical/molecule
import chain first. The last change we need to make for deads is to
transition NIFRID:birnlexRetiredClass to owl:DeprecatedClass.