chemical/molecule tripple shuffle
Moved triples around without moving the location of their deepest point
in the import chain. All higher imports are converted to annotations.

Synthetic enzymes and degredation enzymes are NOT disjoint, this is
trivial if you consider that someday we will made degredation enzymes
from scratch. Probably should deprecate synthetic enzyme entirely since
it is a role.