chemical/molecule import chain, shuffle, and chebi
Major rearrangement of the import chain for the chemical and molecular
ontology files.

All of chebi-bridge has been converted to annotations to reflect the
fact that the classes are defined in the imported file in question
(uberon-bridge still needs to be converted).

Annotations have been lifted to the point in the import chain where
the classes they are annotating are all already imported.

All chem/mol ontologies are now imported via chemical-bridge and then
through molecule-bridge. The design here attempts to minimize the number
of duplicate owl:Class redefinitions and attempts to point people in the
right direction to find where the original classes are defined. This
should help increase the modularity of our ontologies and allow us to
swap parts and execute updates with fewer issues. It should also
simplify signaling about where annotations should be placed. If you want
an annotation on external ontology but it in the bridge file that
directly imports it as an annotation.