NIFGA removal of has_part and has_integral_part
It appears that intention with which these ObjectProperties were
originally used was to convey the idea that an object was the sum of its
parts. HIP was used on unions (and one intersection by mistake) to assert
that the class of the union was part of the whole. Of course this
modelling doesn't make sense because there was not obvious way to infer
that the members of the union were themselves part of the parent class.
I have converted all these cases to has_proper_part (consistent with
usage in the rest of nif) and added the members of the union to an
owl:disjointUnion which seems to have been the original intention based
on some of the editorial notes about the usage of HIP.

@cmungall some of these cases reveal a conflict with Uberon modelling
(for example UBERON:0003002). I will file a bug.