NIFSUB replaced BOF1SNAP:ObjectBoundary
Now using BFO:0000146 2d continuant fiat boundary.
Not including this as a general 1.1->2 translation since there is not
a consistent mapping depending on whether the usage of the boundary
was to refer to a material entity or to some immaterial part.
Given that the subclasses below this are not used anywhere else,
we can make this call without too much trouble.