sparc methods new classes
Added Dataset, DatasetContribution, File, Folder, and Person.

These are classes that are now well established in the existing data
and have distinct meaning. All except Person are subclasses that
specialize sparc:Resource to introduce semantics needed for a variety
of operations acting on those specific subtypes. Person is introduced
becuase Researcher is too narrow, the next commit relaces the use of
Researcher with Person.